+++ refresh/templates/common/main.swift
@@ -8,8 +8,8 @@ do {
 
   HeliumLogger.use(LoggerMessageType.info)
 
-  try initialize()
-  try run()
+  let app = try App()
+  try app.run()
 
 } catch let error {
     Log.error(error.localizedDescription)